rename authRepository to authService
