+++ api/repository/index.ts
@@ -1,4 +1,3 @@
 export { userRepository } from "./user";
 export { orderRepository } from "./order";
 export { eventRepository } from "./event";
-export { authRepository } from "./auth";
