+++ api/services/index.ts
@@ -1,0 +1,1 @@
+export { authService } from "./auth";
